+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # 2.x branch
-## 2.9 branch
-### 2.9.2
+## 2.10 branch
+### 2.10.0
 * now allows to configure and customize via bootstrap the executable commands to
     import and export databases, for each driver;
+* `__exportExecutableWithCompression()` and `_importExecutableWithCompression()`
+    methods provided by the `Driver` class have been removed and incorporated
+    into `_exportExecutable()` and `_importExecutable()`;
 * postgres and sqlite commands are also properly escaped.
 
+## 2.9 branch
 ### 2.9.1
 * all shell arguments are now correctly escaped.